requirements.txt
fitz
scikit-learn~=1.7.0
numpy~=2.3.1
pdfminer.six
sentence_transformers
openai~=1.93.0
python-dotenv~=1.1.1
mcp[cli]~=1.11.0
streamlit~=1.47.1
langchain~=0.3.26
requests~=2.32.4
colorama~=0.4.6
torch~=2.7.1
sentence-transformers~=5.0.0
transformers~=4.53.1
pydantic~=2.11.7
whisper~=1.1.10
pyttsx3~=2.99
crewai~=0.152.0